# wire: Decrease num allocs when decoding headers.