Fixes #357

```
benchmark: BenchmarkDecodeHeaders

old ns/op     new ns/op     delta
-----------------------------------
4517561       3335930       -26.16%

old allocs    new allocs    delta
-----------------------------------
12002         2             -99.98%

old bytes     new bytes     delta
-----------------------------------
561793        417792        -25.63%
```