@davecgh 2017-09-20
Thanks for the update, but the commit still does not conform to the [Model Git Commit Messages](https://github.com/decred/dcrd/blob/master/docs/code_contribution_guidelines.md#ModelGitCommitMessages) section of the [Code Contribution Guidelines](https://github.com/decred/dcrd/blob/master/docs/code_contribution_guidelines.md).

In particular:

* prefix: Short summary of changes (***max 50 chars***)
* More detailed explanatory text, if necessary.  ***Wrap it to about 72***

It also is not reporting the the differences with [benchcmp](https://godoc.org/golang.org/x/tools/cmd/benchcmp) as requested.

For example, this is more in line with the requirements:

```
wire: Decrease num allocs when decoding headers.

The type switch in readElement() does not cover all cases.

Types *uint8, *uint16, *[6]byte and *[32]byte are not taken into account
and, thus, are handled by the slow binary decoder (fall back behavior).

Type cases are added for these 4 types.

These types appears in some BlockHeader attributes.

....results of benchcmp here....
```